@@ -12,7 +12,7 @@
     "dependencies": {
         "@astrojs/mdx": "^2.1.1",
         "@astrojs/rss": "^4.0.5",
-        "@astrojs/svelte": "^5.0.3",
+        "@astrojs/svelte": "^6.0.2",
         "@astrojs/vercel": "^7.3.1",
         "astro": "^4.4.1",
         "hast-util-to-string": "^3.0.0",
@@ -24,7 +24,7 @@
         "sanitize-html": "^2.12.0",
         "satori": "^0.10.13",
         "sharp": "^0.32.6",
-        "svelte": "^5.1.3"
+        "svelte": "^5.3.1"
     },
     "devDependencies": {
         "sass": "^1.71.1"
